+++ src/EventHub/EventHub.Autorest/test/Get-AzEventHubConsumerGroup.Tests.ps1
@@ -15,13 +15,13 @@ if(($null -eq $TestName) -or ($TestName -contains 'Get-AzEventHubConsumerGroup')
 }
 
 Describe 'Get-AzEventHubConsumerGroup' {
-    It 'List' {
+    It 'List'  {
         $listOfConsumerGroup = Get-AzEventHubConsumerGroup -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace -EventHubName $env.eventHub
         $listOfConsumerGroup.Count | Should -Be 1
         $listOfConsumerGroup[0].Name | Should -Be $env.consumerGroup
     }
 
-    It 'Get' {
+    It 'Get'  {
         $consumerGroup = Get-AzEventHubConsumerGroup -ResourceGroupName $env.resourceGroup -NamespaceName $env.namespace -EventHubName $env.eventHub -Name $env.consumerGroup
         $consumerGroup.ResourceGroupName | Should -Be $env.resourceGroup
         $consumerGroup.Name | Should -Be $env.consumerGroup