+++ src/MarketplaceOrdering/exports/ProxyCmdletDefinitions.ps1
@@ -41,7 +41,7 @@ INPUTOBJECT <IMarketplaceOrderingIdentity>: Identity Parameter
   [PublisherId <String>]: Publisher identifier string of image being deployed.
   [SubscriptionId <String>]: The subscription ID that identifies an Azure subscription.
 .Link
-https://docs.microsoft.com/powershell/module/az.marketplaceordering/invoke-azmarketplacesignterms
+https://learn.microsoft.com/powershell/module/az.marketplaceordering/invoke-azmarketplacesignterms
 #>
 function Invoke-AzMarketplaceSignTerms {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.Api202101.IAgreementTerms])]
@@ -232,7 +232,7 @@ INPUTOBJECT <IMarketplaceOrderingIdentity>: Identity Parameter
   [PublisherId <String>]: Publisher identifier string of image being deployed.
   [SubscriptionId <String>]: The subscription ID that identifies an Azure subscription.
 .Link
-https://docs.microsoft.com/powershell/module/az.marketplaceordering/stop-azmarketplaceterms
+https://learn.microsoft.com/powershell/module/az.marketplaceordering/stop-azmarketplaceterms
 #>
 function Stop-AzMarketplaceTerms {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.Api202101.IAgreementTerms])]
@@ -411,7 +411,7 @@ Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.IMarketplaceOrderi
 .Outputs
 Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.Api202101.IAgreementTerms
 .Link
-https://docs.microsoft.com/powershell/module/az.marketplaceordering/get-azmarketplaceterms
+https://learn.microsoft.com/powershell/module/az.marketplaceordering/get-azmarketplaceterms
 #>
 function Get-AzMarketplaceTerms {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.Api202101.IAgreementTerms])]
@@ -615,7 +615,7 @@ TERMS <IAgreementTerms>: Terms properties for provided Publisher/Offer/Plan tupl
   [SystemDataLastModifiedBy <String>]: The identity that last modified the resource.
   [SystemDataLastModifiedByType <CreatedByType?>]: The type of identity that last modified the resource.
 .Link
-https://docs.microsoft.com/powershell/module/az.marketplaceordering/set-azmarketplaceterms
+https://learn.microsoft.com/powershell/module/az.marketplaceordering/set-azmarketplaceterms
 #>
 function Set-AzMarketplaceTerms {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.MarketplaceOrdering.Models.Api202101.IAgreementTerms])]